Move ServerListing to src/types/server-listing.ts
docker/orientation.ts and docker/agent-adapter.ts type-only-imported
ServerListing from session/prompts.ts to share a 4-line {name, description}
interface — a thin docker→session edge that existed solely for this one
shape. Move it to the established small-leaf-type home under src/types/
and update all 6 import sites.

session/prompts.ts keeps a `export type { ServerListing }` re-export so
any caller still pulling the type from the old path keeps working — same
backwards-compat pattern used in PR #222.
