+++ src/session/agent-session.ts
@@ -29,7 +29,8 @@ import type { Sandbox } from '../sandbox/index.js';
 import { SessionNotReadyError, SessionClosedError, BudgetExhaustedError } from '../types/errors.js';
 import { MessageCompactor } from './message-compactor.js';
 import { createCacheStrategy, type PromptCacheStrategy } from './prompt-cache.js';
-import { buildSystemPrompt, type ServerListing } from './prompts.js';
+import { buildSystemPrompt } from './prompts.js';
+import type { ServerListing } from '../types/server-listing.js';
 import { ResourceBudgetTracker } from './resource-budget-tracker.js';
 import { StepLoopDetector } from './step-loop-detector.js';
 import { truncateResult, getResultSizeLimit, formatKB } from './truncate-result.js';
