+++ test/docker-agent-adapter.test.ts
@@ -8,7 +8,7 @@ const claudeCodeAdapter = createClaudeCodeAdapter();
 import { registerAgent, getAgent, listAgents } from '../src/docker/agent-registry.js';
 import { prepareSession, extractAllowedDomains } from '../src/docker/orientation.js';
 import { CONTAINER_WORKSPACE_DIR, type AgentId, type OrientationContext } from '../src/docker/agent-adapter.js';
-import type { ServerListing } from '../src/session/prompts.js';
+import type { ServerListing } from '../src/types/server-listing.js';
 import type { IronCurtainConfig } from '../src/config/types.js';
 
 const sampleServerListings: ServerListing[] = [{ name: 'filesystem', description: 'Read, write, and manage files' }];
