+++ test/help-integration.test.ts
@@ -3,7 +3,8 @@ import { mkdirSync, writeFileSync, rmSync } from 'node:fs';
 import { resolve, dirname } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { Sandbox, type HelpData } from '../src/sandbox/index.js';
-import { buildSystemPrompt, type ServerListing } from '../src/session/prompts.js';
+import { buildSystemPrompt } from '../src/session/prompts.js';
+import type { ServerListing } from '../src/types/server-listing.js';
 import type { IronCurtainConfig } from '../src/config/types.js';
 import type { CompiledPolicyFile, ToolAnnotationsFile } from '../src/pipeline/types.js';
 import { testCompiledPolicy, testToolAnnotations, REAL_TMP } from './fixtures/test-policy.js';
